pyproject.toml
[tool.poetry]
name = "kharon"
version = "0.1.0"
description = "Proxy server with authentication"
authors = ["Dref360 <frederic.branchaud.charron@gmail.com>"]
license = "ApacheV2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.27.1"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
requests = "^2.31.0"
authlib = "^1.3.0"
itsdangerous = "^2.1.2"
google-auth = "^2.28.0"
google-auth-oauthlib = "^1.2.0"
sqlmodel = "^0.0.16"
sshtunnel = "^0.4.0"
platformdirs = "^4.2.2"
names-generator = "^0.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
mypy = "^1.8.0"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
